requirements.txt
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
requests==2.31.0
beautifulsoup4==4.12.2
flask==2.3.3
flask-cors==4.0.0
sqlite3
chromadb==0.4.15
nltk==3.8.1
tqdm==4.66.1
python-dotenv==1.0.0
pydantic==2.4.2
uvicorn==0.23.2
fastapi==0.104.1
